requirements.txt
click==6.7
Flask==0.12.3
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==2.0.0
python-dotenv==0.17.1
pytz==2017.2
requests==2.13.0
Werkzeug==0.12.1
